fix: Integer to Double

+++ src/main/java/com/gavynholt/crud/entity/Location.java
@@ -36,10 +36,10 @@ public class Location {
     private String country;
 
     @Column(name="latitude")
-    private Integer latitude;
+    private Double latitude;
 
     @Column(name="longitude")
-    private Integer longitude;
+    private Double longitude;
 
     @JsonManagedReference
     @OneToMany(mappedBy="location",
@@ -128,19 +128,19 @@ public class Location {
         this.country = country;
     }
 
-    public Integer getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Integer latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
-    public Integer getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Integer longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
